rag-system/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.8
coloredlogs==15.0.1
cryptography==45.0.1
dataclasses-json==0.5.14
distro==1.9.0
emoji==2.14.1
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
Flask==2.0.1
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.267
langchain-community==0.0.10
langchain-core==0.1.23
langdetect==1.0.9
langsmith==0.0.87
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numexpr==2.10.2
numpy==1.26.4
olefile==0.47
onnxruntime==1.22.0
openai==0.28.1
openapi-schema-pydantic==1.2.4
overrides==7.7.0
packaging==23.2
pdf2image==1.17.0
pillow==11.2.1
posthog==4.0.1
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pulsar-client==3.7.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypdf==5.5.0
PyPika==0.48.9
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.5.1
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
unstructured==0.17.2
unstructured-client==0.35.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
webencodings==0.5.1
websockets==15.0.1
Werkzeug==2.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
python-docx==1.1.0
